data_provider/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pynecore-data-provider"
version = "0.1.0"
description = "Data provider plugin for PyneCore"
authors = [
    {name = "PyneCore Contributors"}
]
requires-python = ">=3.10"
dependencies = [
    "pynesys-pynecore",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
ccxt = ["ccxt>=4.0.0"]
capitalcom = ["httpx>=0.24.0"]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]

[project.scripts]
pyne-data = "pynecore_data_provider.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100